+++ connector/src/clients/torrentio.ts
@@ -1,13 +1,13 @@
-import { ZodIssue, z } from "zod";
+import { z } from "zod";
 
 import { DebridCreds, buildDebridPathPart } from "../data/debrid";
-import { CommonError } from "../data/error";
 import log from "../log";
 import { ToFetch } from "../logic/list";
 import { Cache } from "../store/cache";
 import { VERSION } from "../util/version";
 
-import { RequestError, fetchResp, getJSON } from "./http";
+import { fetchResp, getJSON } from "./http";
+import { CommonError } from "./http.types";
 
 const TORRENTIO_HOST = "https://torrentio.strem.fun";
