--- connector/src/data/error.ts
@@ -1,5 +1,0 @@
-import { ZodIssue } from "zod";
-
-import { RequestError } from "../clients/http";
-
-export type CommonError = Error | ZodIssue | RequestError;
